+++ typing-practice-be/src/main/java/com/typingpractice/typing_practice_be/statistics/controller/AdminWordStatisticsController.java
@@ -1,0 +1,38 @@
+package com.typingpractice.typing_practice_be.statistics.controller;
+
+import com.typingpractice.typing_practice_be.common.ApiResponse;
+import com.typingpractice.typing_practice_be.word.statistics.service.GlobalWordStatisticsBatchService;
+import com.typingpractice.typing_practice_be.wordtypingrecord.statistics.service.MemberWordTypingStatsBatchService;
+import com.typingpractice.typing_practice_be.wordtypingrecord.statistics.service.WordTypingStatsBatchService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/admin/stats")
+public class AdminWordStatisticsController {
+  private final GlobalWordStatisticsBatchService globalWordStatisticsBatchService;
+  private final WordTypingStatsBatchService wordTypingStatsBatchService;
+  private final MemberWordTypingStatsBatchService memberWordTypingStatsBatchService;
+
+  @PostMapping("/global-word/recalculate")
+  public ApiResponse<Void> recalculateGlobalWordStats() {
+    globalWordStatisticsBatchService.runManualRecalculation();
+    return ApiResponse.ok(null);
+  }
+
+  @PostMapping("/word-typing/recalculate")
+  public ApiResponse<Void> recalculateWordTypingStats() {
+    wordTypingStatsBatchService.runManualRecalculation();
+    return ApiResponse.ok(null);
+  }
+
+  @PostMapping("/member-word-typing/recalculate")
+  public ApiResponse<Void> recalculateMemberWordTypingStats() {
+    memberWordTypingStatsBatchService.runManualRecalculation();
+
+    return ApiResponse.ok(null);
+  }
+}
